Add files via upload
Kinetoplastid membrane protein-11 (KMP-11), of size 11 kDa is a small, conserved flagellar protein amongst all kinetoplastid parasites. It is highly expressed throughout the Leishmania life cycle and is a potential leishmaniasis vaccine candidate. This study focuses on the interactions between the KMP-11 protein of the Leishmania parasite and artificial lipid membranes using an unsaturated 1,2-dioleoyl-sn-glycero-3-phosphocholine (DOPC) and saturated 1,2-didecanoyl-sn-glycero-3-phosphocholine (DPPC) lipid. LUVs (Large unilamellar vesicles) were made with DOPC and SUVs (Small unilamellar vesicles) were made with DPPC lipids respectively. Calcein release assay was performed and as observed from the fluorometric experiments, the fluorescence intensity was seen to increase with increasing concentrations of KMP-11 in both calcein infused LUVs and SUVs. This assay, which evaluated the effect of KMP-11 on vesicle disruption, indicated that the protein might have pore-forming capabilities, causing the release and fluorescence of calcein encapsulated within the vesicles.